🐛 fix: six bugs from a fourth audit, most of them dtype
The audit validated all five focus items on the previous performance commit
-- the inlined `K2e` is bit-identical over 6,300 points, `K1e` leaks nothing
through its `where` in any mode, and `2/|z|`'s sign genuinely cannot matter
because every negative `z` is already `nan` from `log(z)`. What it found
instead was a family of dtype defects nothing had ever exercised.

**`grad(K2)` raised outright on float16/bfloat16.** `_K2_jvp` narrowed its
primal with `_cast_like` but left the tangent at the internal width, so JAX
rejected the rule ("must produce primal and tangent outputs with corresponding
... dtypes"). It was the only rule that cast its primal, which is why it was
the only one that raised -- the other five cast neither, and so quietly handed
`jax.jvp` a wider primal than the plain call returns. Every rule now casts
both.

**float32 input returned float64** from all six K functions, `zeta` and
`spence`, whenever x64 was on. `jnp.arange(1.0, n)` defaults to float64 there,
so the series constants promoted the whole computation; `zeta` widened through
the float64 Bernoulli table. This is the hazard the narrow-dtype test was
written for -- a widened output breaks a `lax.scan` carry -- one dtype up, and
invisible because every test runs in float64.

**`Li(2, float32(0.5))` died inside `lax.fori_loop`**, because the carry was
seeded at the argument's width and the body widened it through the same
Bernoulli table. The error named neither `Li` nor the dtype.

**`grad(grad(K1))` and `grad(grad(K2))` were 7.2e-4 low from z ~ 694** -- the
subnormal-flush bug the first derivative was fixed for, one order up. The
first derivative was an *expression* inside the JVP, so differentiating it
again formed `d(1/z) * K1 * e^-z`, about 4e-312 at z = 700. `K1'` and `K2'`
are now named functions with their own analytic rules
(K1'' = K1 + K0/z + 2K1/z^2, K2'' = K0 + 3K1/z + 6K2/z^2, both verified
exact against mpmath), so the second derivative is summed scaled as well. Now
2e-16 across the tail.

**The accuracy figures were understated ~1.8x.** A 400,000-point sweep of
[8.90, 9.00] puts the peak at z = 8.9932: 2.2e-7 for `K0`, 2.0e-7 for `K1`,
1.4e-7 for `K2` against mpmath. My earlier sweep sampled logarithmically
around z = 9 and stepped over it. Corrected in six docstrings, the docs, both
guidance files and three test docstrings -- including one in
`test_scipy_parity.py` still reading `8e-8`, which the previous commit was
supposed to have purged.

Also documented: the float32 ceiling at z ~ 85.3 (where K0 drops below
float32's smallest normal -- the same wall as 705.5, one dtype down), and that
`grad(grad)` at exactly z = 0 is `inf` for the unscaled family but `nan` for
the scaled one, since `2/|z|` has derivative `sign(0) = 0`.

Two parity gates had orders of dead slack: `zeta` asserted rtol 1e-12 against
a measured 6.7e-16, and `Li` 1e-11 against 1.3e-14. Both would have passed a
100x regression; now 5e-15 and 1e-13.

+++ src/spexial/_src/zeta.py
@@ -91,6 +91,9 @@ def zeta(n: RealArrayLike, /) -> AnyArray:
     1.0
 
     """
+    # `* 1.0` promotes integers; the Bernoulli table is float64 by construction
+    # (exact `Fraction` arithmetic), so it is cast down to `n_arr`'s dtype below
+    # rather than being allowed to widen a float32 argument to float64.
     n_arr = jnp.asarray(n) * 1.0
     positive = n_arr > 0
     k = -n_arr  # zeta(-k)
@@ -114,7 +117,7 @@ def zeta(n: RealArrayLike, /) -> AnyArray:
     reflected = jnp.where(
         ~is_integer | (k_round + 1.0 > ORDER),
         jnp.nan,
-        sign * bernoulli_numbers()[index] / denom,
+        sign * bernoulli_numbers().astype(n_arr.dtype)[index] / denom,
     )
 
     # `_hurwitz_zeta` is `nan` for n above ~1e15, and is exactly 1.0 for every n
